WARNING: INCOMPATIBLE WITH PREVIOUS VERSIONS. read more...
this is the first time ledBlaster is not downwards compatible.
the reason is a rather major change in the input protocol made for using the blaster

the old key=value solution is now not longer functional.
it was replaced with a new:
[char of applyMode]key:value;key:value ...
protocol.

further clarifications:
[char of applyMode] means:
	one single character specifying what to do e.g.
	s -> set new brightness
	f -> fade new brightness
	new config parameters to come

benefits:
- change all values in one line
- or only one value in one line (no overhead)
- no need to change fadetime before every set (live manipulation mode)
- this protocol is compatible with the esp-led protocol which is currently under development

todos:
- add configuration options
----> the long term goal is to abolish the .csv files and replace them with json files.
led blaster is then live configured via nodejs server (currently called forward.js, it gets a branch once needed)
nodejs is master, led-blaster is slave

also done
launch forward.js while launching led-blaster