introducing live initialization
fifo.cpp and led.cpp are the most powerful sources atm.
the led objects are now created by fifo commands and are no longer limited to the old csv files.
nodejs here we come! ;)
many old things are now commented and some cleanup should be done.. maybe file.cpp is obsolete.

known bugs:
sometimes it flickers on startup (after the init lines have been read, I haven't been able to reproduce it. :/

fixes:
fixed error where lines weren't read correctly and caused stoi to fail.
this results in the condition that a semicolon ; has to be at the end of EVERY STATEMENT, otherwise it won't be
processed. It is maybe not optimal but definitely better than a program which is crashing.